Reorganize classes and modules
