+++ OTAnalytics/adapter_ui/flow_adapter.py
@@ -1,9 +1,5 @@
 from abc import ABC, abstractmethod
 
-from customtkinter import CTkFrame
-
-from OTAnalytics.adapter_ui.abstract_frame import AbstractCTkFrame
-from OTAnalytics.adapter_ui.helpers import WidgetPositionProvider
 from OTAnalytics.domain.section import COORDINATES, Section
 
 
@@ -56,9 +52,3 @@ class GeometricCenterCalculator(SectionRefPointCalculator):
         center_x = total_x / num_points
         center_y = total_y / num_points
         return center_x, center_y
-
-
-class AbstractFrameFlows(AbstractCTkFrame, WidgetPositionProvider, CTkFrame):
-    @abstractmethod
-    def introduce_to_viewmodel(self) -> None:
-        pass
