+++ OTAnalytics/plugin_ui/customtkinter_gui/abstract_ctk_frame_flows.py
@@ -1,0 +1,12 @@
+from abc import abstractmethod
+
+from customtkinter import CTkFrame
+
+from OTAnalytics.adapter_ui.abstract_frame import AbstractCTkFrame
+from OTAnalytics.adapter_ui.helpers import WidgetPositionProvider
+
+
+class AbstractFrameFlows(AbstractCTkFrame, WidgetPositionProvider, CTkFrame):
+    @abstractmethod
+    def introduce_to_viewmodel(self) -> None:
+        raise NotImplementedError
