+++ OTAnalytics/plugin_ui/customtkinter_gui/abstract_ctk_frame_sections.py
