+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_flows.py
@@ -4,9 +4,11 @@ from typing import Any
 
 from customtkinter import CTkButton, CTkFrame, CTkScrollbar
 
-from OTAnalytics.adapter_ui.abstract_frame_flows import AbstractFrameFlows
 from OTAnalytics.adapter_ui.view_model import ViewModel
 from OTAnalytics.domain.flow import Flow
+from OTAnalytics.plugin_ui.customtkinter_gui.abstract_ctk_frame_flows import (
+    AbstractFrameFlows,
+)
 from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 from OTAnalytics.plugin_ui.customtkinter_gui.helpers import get_widget_position
 from OTAnalytics.plugin_ui.customtkinter_gui.treeview_template import (
