+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_sections.py
@@ -5,9 +5,11 @@ from typing import Any
 
 from customtkinter import CTkButton, CTkFrame, CTkScrollbar
 
-from OTAnalytics.adapter_ui.abstract_frame_sections import AbstractFrameSections
 from OTAnalytics.adapter_ui.view_model import ViewModel
 from OTAnalytics.domain.section import Section
+from OTAnalytics.plugin_ui.customtkinter_gui.abstract_ctk_frame_sections import (
+    AbstractFrameSections,
+)
 from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 from OTAnalytics.plugin_ui.customtkinter_gui.helpers import get_widget_position
 from OTAnalytics.plugin_ui.customtkinter_gui.treeview_template import (
